requirements.txt
torch
numpy
huggingface_hub
dataclasses; python_version < "3.7"
python-dotenv
datasets
matplotlib